backoffice/requirements.txt
Django
flake8
mysqlclient
pylint
djangorestframework
djangorestframework-jwt
django-rest-swagger
